refactor(auth): extract useAuth hook from AuthProvider and update imports
- Move useAuth hook to separate file for better code organization
- Update all imports to use the new hook location
- Optimize AuthProvider with useMemo for context value
- Simplify RoleRedirect component by removing useEffect
