+++ src/shared/components/router/ProtectedRoute.tsx
@@ -2,7 +2,7 @@
 
 import React from "react";
 import { Navigate, useLocation } from "react-router-dom";
-import { useAuth } from "../../../features/auth/context/AuthProvider";
+import { useAuth } from "../../../features/auth/hooks/useAuth";
 import { CircularProgress, Box } from "@mui/material";
 
 interface ProtectedRouteProps {
